fixed category parent id type

+++ database/migrations/2021_06_11_105535_add_parent_column_to_categories_table.php
@@ -14,7 +14,7 @@ class AddParentColumnToCategoriesTable extends Migration
     public function up()
     {
         Schema::table('categories', function (Blueprint $table) {
-            $table->integer('parent_id')->unsigned()->nullable();
+            $table->unsignedBigInteger('parent_id')->unsigned()->nullable();
 
             $table->foreign('parent_id')->references('id')->on('categories');
         });
